Reduce code duplication in build makefiles
The makefiles for BUILD_EXECUTABLE and BUILD_SHARED_LIBRARY share a lot
of code. This patch deduplicates the common code into
prebuilts.internal.mk.

There is one notable semantic change now. If LOCAL_PREBUILT_MODULE_FILE
was already set by the caller, we now do not use our own prebuilts
system anymore.

Another notable change is that cached file names have changed to include
the git revision for the source directory now. That makes it easier to
track file and revision in one spot. There is no need to maintain a
separate file for the Git revision anymore. This was flakey from the
start. The cache file and revision file could have gone out of sync.
And it greatly reduces the code.