Rework organization of binary cache
Cached files are organized in a different way now. Instead of
hard-coding a directory per device/build-variant, we now choose all
information necessary dynamically. The goal is to share a large part of
the cache across as many devices and build variants as possible.

We also try to be smart about special build flags that a binary can
depend upon, like build variant or build type. There are probably some
missing, but it's a good start.