More meaningful statistics
Count all modules that are parsed by prebuilts.internals.mk. Compare it
to the number of modules we find prebuilts for. That number is way more
meaningful than all files on the prebuilts cache, because the cache
might contain files for different devices or even stale files for older
versions.