Rework obtaining git revisions for all modules
An Android project may consist of multiple repositories. A module in an
upper-level directory structure normally depends on modules in deeper
directories. One example would be libwebviewchromium from
external/chromium_org/.

So not only the git revision of the module path itself is of interest,
but also the revisions of repositories in deeper directory structures.
The same holds true for the dirty state of any of the git repositories.